+++ pg_documentdb/src/commands/commands_common.c
@@ -197,13 +197,15 @@ FindShardKeyValueForDocumentId(MongoCollection *collection, const bson_value_t *
 	{
 		/* set the variableSpec */
 		argTypes[1] = bsonTypeId;
-		argValues[1] = applyVariableSpec ? PointerGetDatum(variableSpecBson) : (Datum) 0;
-		argNulls[1] = applyVariableSpec ? ' ' : 'n';
+		argValues[1] = applyVariableSpec ? PointerGetDatum(variableSpecBson) :
+					   PointerGetDatum(PgbsonInitEmpty());
+		argNulls[1] = ' ';
 
 		/* set the collation string */
 		argTypes[2] = TEXTOID;
-		argValues[2] = applyCollation ? CStringGetTextDatum(collationString) : (Datum) 0;
-		argNulls[2] = applyCollation ? ' ' : 'n';
+		argValues[2] = applyCollation ? CStringGetTextDatum(collationString) :
+					   CStringGetTextDatum("");
+		argNulls[2] = ' ';
 	}
 
 	/* set _id filter */